feat(evals): add LM-judge eval harness for the plan-generator
Establishes the eval layer flagged by the "New SDLC / agentic engineering"
benchmark: tests verify the code, this verifies the AI output itself.

- Extract pure prompt/schema/normalize logic into lib/plan/plan-core.ts so the
  eval exercises the EXACT production generator (no prompt drift); route imports
  it, behavior unchanged (route.test.ts + planGenerator.test.ts still pass).
- evals/plan-generator/: 12-persona golden set, deterministic safety checks
  (schema, week count, distance sanity, weekly load cap, no-hard-sessions for
  recovery/mindful), and a gpt-4o-mini LM-judge rubric.
- Free offline checks run in normal CI; paid LM-judge gated behind RUN_LIVE_EVAL
  via npm run eval:plan + nightly/pre-release GitHub Action.
- Gate: zero safety-critical failures + judge pass rate >= 0.85.

Baseline (2026-06-26): 0/12 pass. Surfaced real findings recorded in
BASELINE.md: beginner/mindful plans get hard sessions, 12-week plans truncate
past maxOutputTokens and silently fall back, and some plans violate the 10%
load rule. Thresholds intentionally left at target, not loosened to pass.
