+++ v0/.gitignore
@@ -26,3 +26,6 @@ yarn-error.log*
 *.tsbuildinfo
 next-env.d.ts
 .env*.local
+
+# Eval run artifacts (regenerated each run)
+evals/**/report.json
